+++ features/screen/main_navigation_contents/src/main/java/com/season/winter/main_navigation_contents/fragments/navigation/home/HomeFragment.kt
@@ -2,13 +2,11 @@ package com.season.winter.main_navigation_contents.fragments.navigation.home
 
 import android.util.Log
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView.LayoutManager
 import com.season.winter.common.extention.coroutine.repeatOnLifecycle
 import com.season.winter.main_navigation_contents.R
 import com.season.winter.main_navigation_contents.databinding.FragmentHomeBinding
 import com.season.winter.main_navigation_contents.fragments.base.BaseNavigationFragment
 import com.season.winter.main_navigation_contents.fragments.navigation.home.recyclerView.adapter.HomeRecyclerViewAdapter
-import com.season.winter.ui.model.fragment.home.HomeItemType
 
 class HomeFragment: BaseNavigationFragment<FragmentHomeBinding>(R.layout.fragment_home) {
