+++ features/screen/main_navigation_contents/src/main/java/com/season/winter/main_navigation_contents/fragments/navigation/home/recyclerView/adapter/HomeDiffCallback.kt
@@ -1,7 +1,7 @@
 package com.season.winter.main_navigation_contents.fragments.navigation.home.recyclerView.adapter
 
 import androidx.recyclerview.widget.DiffUtil
-import com.season.winter.ui.model.fragment.home.HomeItem
+import com.season.winter.liquor.dummy.model.HomeItem
 
 class HomeDiffCallback: DiffUtil.ItemCallback<HomeItem>() {
