+++ features/screen/main_navigation_contents/src/main/res/layout/item_liquor_grouping.xml
@@ -6,7 +6,7 @@
     <data>
 
 <!--        <import type="com.season.winter.liquor.liquorInfo.LiquorInfo" />-->
-        <import type="com.season.winter.ui.model.fragment.home.HomeItem" />
+        <import type="com.season.winter.liquor.dummy.model.HomeItem" />
         <import type="com.season.winter.main_navigation_contents.fragments.navigation.home.recyclerView.adapter.HomeRecyclerViewAdapter" />
         <import type="com.season.winter.main_navigation_contents.fragments.navigation.home.recyclerView.viewHolder.HomeViewHolder" />
         <import type="com.season.winter.main_navigation_contents.ui.recyclerView.adapter.MiniLiquorAdapter" />
@@ -17,7 +17,7 @@
 
         <variable
             name="home_item"
-            type="com.season.winter.ui.model.fragment.home.HomeItem"/>
+            type="com.season.winter.liquor.dummy.model.HomeItem"/>
 
         <variable
             name="adapter"
