+++ repository/liquor/src/main/java/com/season/winter/liquor/dummy/dataSet/liquorInfo/KoreaLiquorInfo.kt
@@ -7,10 +7,11 @@ import com.season.winter.liquor.liquorInfo.WhiskyType
 import com.season.winter.liquor.brand.BrandInfo
 import com.season.winter.liquor.content.FeatureData
 import com.season.winter.liquor.content.FeatureType
+import com.season.winter.liquor.dummy.dummy.getKoreanLiquorIdentify
 import com.season.winter.liquor.liquorInfo.LiquorStatus
 
 val kiOneBatch1 = LiquorInfo(
-    identify = com.season.winter.liquor.dummy.getKoreanLiquorIdentify(0),
+    identify = getKoreanLiquorIdentify(0),
     nameKo = "기원 배치 1",
     nameEn = "KI ONE Batch 1",
     wishCount = 0,
@@ -69,7 +70,7 @@ val kiOneBatch1 = LiquorInfo(
 )
 
 val kiOneBatch2 = LiquorInfo(
-    identify = com.season.winter.liquor.dummy.getKoreanLiquorIdentify(1),
+    identify = getKoreanLiquorIdentify(1),
     nameKo = "기원 배치 2",
     nameEn = "KI ONE Batch 2",
     wishCount = 0,
