+++ repository/liquor/src/main/java/com/season/winter/liquor/dummy/dummy/IdentifyManage.kt
@@ -1,4 +1,4 @@
-package com.season.winter.liquor.dummy
+package com.season.winter.liquor.dummy.dummy
 
 const val AmericanLiquorIdentify = 10
 const val KoreanLiquorIdentify = 20
