+++ repository/liquor/src/main/java/com/season/winter/liquor/dummy/dummy/LiquorInfoDummyGenerator.kt
@@ -1,4 +1,4 @@
-package com.season.winter.liquor.dummy
+package com.season.winter.liquor.dummy.dummy
 
 import com.season.winter.liquor.liquorInfo.LiquorInfo
 import com.season.winter.liquor.dummy.dataSet.liquorInfo.americanLiquorInfoList
@@ -7,7 +7,6 @@ import com.season.winter.liquor.dummy.dataSet.liquorInfo.scotchLiquorInfoList
 import com.season.winter.storage.ImageFireStorageInstance
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.SharedFlow
