+++ repository/liquor/src/main/java/com/season/winter/liquor/dummy/dummy/LiquorShopDummyData.kt
@@ -1,0 +1,2 @@
+package com.season.winter.liquor.dummy.dummy
+
