+++ repository/screen/src/main/java/com/season/winter/screen/fragment/navigationMain/home/di/HomeNavigationService.kt
@@ -1,6 +1,6 @@
 package com.season.winter.screen.fragment.navigationMain.home.di
 
-import com.season.winter.ui.model.fragment.home.HomeItem
+import com.season.winter.liquor.dummy.model.HomeItem
 import kotlinx.coroutines.flow.Flow
 
 interface HomeNavigationService {
